Moved logic for game into separate file.
